refactor: moving bookmark logic from UserController to BookmarkController
