+++ itachallenge-user/src/main/java/com/itachallenge/user/controller/UserController.java
@@ -3,7 +3,6 @@ package com.itachallenge.user.controller;
 import com.itachallenge.user.annotations.ValidGithubUsername;
 import com.itachallenge.user.document.UserDocument;
 import com.itachallenge.user.service.UserService;
-import com.itachallenge.userinteraction.service.bookmark.BookmarkService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
@@ -12,7 +11,6 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -28,11 +26,9 @@ public class UserController {
     public static final String X_GITHUB_USERNAME ="X-Github-Username";
 
     private final UserService userService;
-    private final BookmarkService bookmarkService;
 
-    public UserController(UserService userService, BookmarkService bookmarkService) {
+    public UserController(UserService userService) {
         this.userService = userService;
-        this.bookmarkService = bookmarkService;
     }
 
     @GetMapping(value = "/test")
@@ -86,120 +82,4 @@ public class UserController {
                             .body(user);
                 });
     }
-
-
-    @Operation(
-            summary = "Add Challenge to User Bookmark Challenges",
-            description = "Adds challenge to user Bookmarks",
-            parameters = {
-                    @Parameter(
-                            name = "userId",
-                            description = "User ID",
-                            required = true,
-                            in = ParameterIn.PATH
-                    ),
-                    @Parameter(
-                            name = "challengeId",
-                            description = "Challenge ID",
-                            required = true,
-                            in = ParameterIn.PATH
-                    )
-            },
-            responses = {
-                    @ApiResponse(
-                            responseCode = "200",
-                            description = "Challenge is already in bookmarks",
-                            content = @Content(mediaType = "application/json")
-                    ),
-                    @ApiResponse(
-                            responseCode = "201",
-                            description = "Challenge added to bookmarks",
-                            content = @Content(mediaType = "application/json")
-                    ),
-                    @ApiResponse(
-                            responseCode = "400",
-                            description = "Bad Request. The provided IDs have a bad format",
-                            content = @Content(mediaType = "application/json")
-                    ),
-                    @ApiResponse(
-                            responseCode = "404",
-                            description = "Not found. No user is found with the provided user id.",
-                            content = @Content(mediaType = "application/json")
-                    ),
-                    @ApiResponse(
-                            responseCode = "500",
-                            description = "Internal server error. An unexpected error occurred.",
-                            content = @Content(mediaType = "application/json")
-                    )
-            }
-    )
-
-    @PostMapping("/users/{userId}/bookmarks/{challengeId}")
-    public Mono<ResponseEntity<Boolean>> addChallengeToBookmarks(@PathVariable String userId, @PathVariable String challengeId) {
-        return bookmarkService.addChallengeToBookmarks(userId, challengeId)
-                .map(added -> {
-                    if (Boolean.TRUE.equals(added)) {
-                        log.info("Challenge '{}' added to user '{}' bookmarks", challengeId, userId);
-                        return ResponseEntity.status(HttpStatus.CREATED).body(true);
-                    } else {
-                        log.info("User's '{}' bookmarks already contain Challenge '{}'", userId, challengeId);
-                        return ResponseEntity.ok().body(false);
-                    }
-                });
-    }
-
-    @Operation(
-            summary = "Delete Challenge from User Bookmark Challenges",
-            description = "Deletes challenge from user bookmarks",
-            parameters = {
-                    @Parameter(
-                            name = "userId",
-                            description = "User ID",
-                            required = true,
-                            in = ParameterIn.PATH
-                    ),
-                    @Parameter(
-                            name = "challengeId",
-                            description = "Challenge ID",
-                            required = true,
-                            in = ParameterIn.PATH
-                    )
-            },
-            responses = {
-                    @ApiResponse(
-                            responseCode = "200",
-                            description = "Challenge deleted from bookmarks or was not in bookmarks",
-                            content = @Content(mediaType = "application/json")
-                    ),
-                    @ApiResponse(
-                            responseCode = "400",
-                            description = "Bad Request. The provided IDs have a bad format",
-                            content = @Content(mediaType = "application/json")
-                    ),
-                    @ApiResponse(
-                            responseCode = "404",
-                            description = "Not found. No user is found with the provided user id.",
-                            content = @Content(mediaType = "application/json")
-                    ),
-                    @ApiResponse(
-                            responseCode = "500",
-                            description = "Internal server error. An unexpected error occurred.",
-                            content = @Content(mediaType = "application/json")
-                    )
-            }
-    )
-    @DeleteMapping("/users/{userId}/bookmarks/{challengeId}")
-    public Mono<ResponseEntity<Boolean>> deleteChallengeFromBookmarks(@PathVariable String userId, @PathVariable String challengeId) {
-        return bookmarkService.deleteChallengeFromBookmarks(userId, challengeId)
-                .map(deleted -> {
-                    if (Boolean.TRUE.equals(deleted)) {
-                        log.info("Challenge '{}' deleted from user '{}' bookmarks", challengeId, userId);
-                        return ResponseEntity.ok().body(true);
-                    } else {
-                        log.info("No change, User's '{}' bookmarks doesn't contain Challenge '{}'", userId, challengeId);
-                        return ResponseEntity.ok().body(false);
-                    }
-                });
-    }
-
 }
