+++ itachallenge-user/src/test/java/com/itachallenge/user/controller/UserControllerTest.java
@@ -2,11 +2,9 @@ package com.itachallenge.user.controller;
 
 import com.itachallenge.user.document.UserDocument;
 import com.itachallenge.user.document.enums.Role;
-import com.itachallenge.common.exception.BadUUIDException;
 import com.itachallenge.common.exception.NotFoundException;
 import com.itachallenge.common.exception.UserGlobalExceptionHandler;
 import com.itachallenge.user.service.UserService;
-import com.itachallenge.userinteraction.service.bookmark.BookmarkService;
 import org.junit.jupiter.api.*;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
@@ -23,9 +21,6 @@ class UserControllerTest {
 
     @Mock
     private UserService userService;
-
-    @Mock
-    private BookmarkService bookmarkService;
 
     @InjectMocks
     private UserController userController;
@@ -103,165 +98,4 @@ class UserControllerTest {
 
         verify(userService, times(1)).getUser(githubUsername);
     }
-
-    @Test
-    void addToBookmarks_WhenAdded_Returns201() {
-        String userId = UUID.randomUUID().toString();
-        String challengeId = UUID.randomUUID().toString();
-        when(bookmarkService.addChallengeToBookmarks(userId, challengeId))
-                .thenReturn(Mono.just(true));
-
-        webTestClient.post()
-                .uri("/itachallenge/api/v1/user/users/" + userId + "/bookmarks/" + challengeId)
-                .exchange()
-                .expectStatus().isEqualTo(HttpStatus.CREATED)
-                .expectBody(Boolean.class).isEqualTo(true);
-
-        verify(bookmarkService, times(1)).addChallengeToBookmarks(userId, challengeId);
-    }
-
-    @Test
-    void addToBookmarks_WhenAlreadyInBookmarks_Returns200() {
-        String userId = UUID.randomUUID().toString();
-        String challengeId = UUID.randomUUID().toString();
-        when(bookmarkService.addChallengeToBookmarks(userId, challengeId))
-                .thenReturn(Mono.just(false));
-
-        webTestClient.post()
-                .uri("/itachallenge/api/v1/user/users/" + userId + "/bookmarks/" + challengeId)
-                .exchange()
-                .expectStatus().isOk()
-                .expectBody(Boolean.class).isEqualTo(false);
-
-        verify(bookmarkService, times(1)).addChallengeToBookmarks(userId, challengeId);
-    }
-
-    @Test
-    void addToBookmarks_WhenUserNotExists_Returns404() {
-        String userId = UUID.randomUUID().toString();
-        String challengeId = UUID.randomUUID().toString();
-        when(bookmarkService.addChallengeToBookmarks(userId, challengeId))
-                .thenReturn(Mono.error(new NotFoundException("User not found")));
-
-        webTestClient.post()
-                .uri("/itachallenge/api/v1/user/users/" + userId + "/bookmarks/" + challengeId)
-                .exchange()
-                .expectStatus().isEqualTo(HttpStatus.NOT_FOUND)
-                .expectBody(String.class).isEqualTo("User not found");
-
-        verify(bookmarkService, times(1)).addChallengeToBookmarks(userId, challengeId);
-    }
-
-    @Test
-    void addToBookmarks_WhenBadFormattedId_Returns400() {
-        String userId = "invalidUuid";
-        String challengeId = "invalidUUid";
-        when(bookmarkService.addChallengeToBookmarks(userId, challengeId))
-                .thenReturn(Mono.error(new BadUUIDException("Error message")));
-
-        webTestClient.post()
-                .uri("/itachallenge/api/v1/user/users/" + userId + "/bookmarks/" + challengeId)
-                .exchange()
-                .expectStatus().isEqualTo(HttpStatus.BAD_REQUEST)
-                .expectBody(String.class).isEqualTo("The provided IDs are not valid.");
-
-        verify(bookmarkService, times(1)).addChallengeToBookmarks(userId, challengeId);
-    }
-
-    @Test
-    void addToBookmarks_WhenUnexpectedError_Returns500() {
-        String userId = UUID.randomUUID().toString();
-        String challengeId = UUID.randomUUID().toString();
-        when(bookmarkService.addChallengeToBookmarks(userId, challengeId))
-                .thenReturn(Mono.error(new Exception()));
-
-        webTestClient.post()
-                .uri("/itachallenge/api/v1/user/users/" + userId + "/bookmarks/" + challengeId)
-                .exchange()
-                .expectStatus().isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR)
-                .expectBody(String.class).isEqualTo("Unexpected error happened.");
-
-        verify(bookmarkService, times(1)).addChallengeToBookmarks(userId, challengeId);
-    }
-
-    @Test
-    void deleteFromBookmarks_WhenDeleted_Returns200() {
-        String userId = UUID.randomUUID().toString();
-        String challengeId = UUID.randomUUID().toString();
-        when(bookmarkService.deleteChallengeFromBookmarks(userId, challengeId))
-                .thenReturn(Mono.just(true));
-
-        webTestClient.delete()
-                .uri("/itachallenge/api/v1/user/users/" + userId + "/bookmarks/" + challengeId)
-                .exchange()
-                .expectStatus().isEqualTo(HttpStatus.OK)
-                .expectBody(Boolean.class).isEqualTo(true);
-
-        verify(bookmarkService, times(1)).deleteChallengeFromBookmarks(userId, challengeId);
-    }
-
-    @Test
-    void deleteFromBookmarks_WhenNotInBookmarks_Returns200() {
-        String userId = UUID.randomUUID().toString();
-        String challengeId = UUID.randomUUID().toString();
-        when(bookmarkService.deleteChallengeFromBookmarks(userId, challengeId))
-                .thenReturn(Mono.just(false));
-
-        webTestClient.delete()
-                .uri("/itachallenge/api/v1/user/users/" + userId + "/bookmarks/" + challengeId)
-                .exchange()
-                .expectStatus().isOk()
-                .expectBody(Boolean.class).isEqualTo(false);
-
-        verify(bookmarkService, times(1)).deleteChallengeFromBookmarks(userId, challengeId);
-    }
-
-    @Test
-    void deleteFromBookmarks_WhenUserNotExists_Returns404() {
-        String userId = UUID.randomUUID().toString();
-        String challengeId = UUID.randomUUID().toString();
-        when(bookmarkService.deleteChallengeFromBookmarks(userId, challengeId))
-                .thenReturn(Mono.error(new NotFoundException("User not found")));
-
-        webTestClient.delete()
-                .uri("/itachallenge/api/v1/user/users/" + userId + "/bookmarks/" + challengeId)
-                .exchange()
-                .expectStatus().isEqualTo(HttpStatus.NOT_FOUND)
-                .expectBody(String.class).isEqualTo("User not found");
-
-        verify(bookmarkService, times(1)).deleteChallengeFromBookmarks(userId, challengeId);
-    }
-
-    @Test
-    void deleteFromBookmarks_WhenBadFormattedId_Returns404() {
-        String userId = "invalidUuid";
-        String challengeId = "invalidUUid";
-        when(bookmarkService.deleteChallengeFromBookmarks(userId, challengeId))
-                .thenReturn(Mono.error(new BadUUIDException("Error message")));
-
-        webTestClient.delete()
-                .uri("/itachallenge/api/v1/user/users/" + userId + "/bookmarks/" + challengeId)
-                .exchange()
-                .expectStatus().isEqualTo(HttpStatus.BAD_REQUEST)
-                .expectBody(String.class).isEqualTo("The provided IDs are not valid.");
-
-        verify(bookmarkService, times(1)).deleteChallengeFromBookmarks(userId, challengeId);
-    }
-
-    @Test
-    void deleteFromBookmarks_WhenUnexpectedError_Returns500() {
-        String userId = UUID.randomUUID().toString();
-        String challengeId = UUID.randomUUID().toString();
-        when(bookmarkService.deleteChallengeFromBookmarks(userId, challengeId))
-                .thenReturn(Mono.error(new Exception()));
-
-        webTestClient.delete()
-                .uri("/itachallenge/api/v1/user/users/" + userId + "/bookmarks/" + challengeId)
-                .exchange()
-                .expectStatus().isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR)
-                .expectBody(String.class).isEqualTo("Unexpected error happened.");
-
-        verify(bookmarkService, times(1)).deleteChallengeFromBookmarks(userId, challengeId);
-    }
-
 }
